+++ crates/bevy_eventlistener_core/src/event_dispatcher.rs
@@ -1,7 +1,6 @@
-use bevy::{
-    prelude::*,
-    utils::{HashMap, HashSet},
-};
+use bevy_ecs::prelude::*;
+use bevy_hierarchy::Parent;
+use bevy_utils::{HashMap, HashSet};
 
 use crate::{
     callbacks::{CallbackSystem, ListenerInput},